adding container component modal
