+++ src/components/ModalContainer/ModalContainer.module.scss
@@ -1,0 +1,31 @@
+.ModalBackdrop {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(20, 18, 18, 0.89);
+  z-index: 999;
+}
+
+.ModalContainer {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  max-height: 100%;
+  width: 540px;
+  background: #ffffff;
+  background-position: 50% 50%;
+  border: none;
+  border-radius: 20px;
+  font-family: Circe;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 1.5;
+}
+@media screen and (max-width: 767px) {
+  .ModalContainer {
+    width: 320px;
+  }
+}
