--- src/components/ModalLogout/ModalLogout.module.scss
@@ -1,62 +1,0 @@
-.ModalBackdrop {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(20, 18, 18, 0.89);
-  z-index: 999;
-}
-
-.ModalContent {
-  position: absolute;
-
-  padding: 20px;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  max-height: 100%;
-  width: 280px;
-  background-image: linear-gradient(to right, #e7eaf2 35%, #ffd8d0 80%);
-  background-position: 50% 50%;
-  border: none;
-  border-radius: 8px;
-  font-family: Circe;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 1.5;
-}
-
-.buttonLogoutYN {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100px;
-  height: 40px;
-  padding: 0px;
-  border: none;
-  border-radius: 8px;
-
-  background-color: #6e78e8;
-  cursor: pointer;
-  color: #ff6596;
-  font-size: 24px;
-  line-height: 1.33;
-  box-shadow: 0px 3px 1px -1px rgba(0, 0, 0, 0.2),
-    0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12);
-  transition: transform 150ms cubic-bezier(0.4, 0, 0.2, 1);
-  &:hover {
-    background-color: #4a56e2;
-    transform: scale(1.02);
-  }
-  &:focus {
-    background-color: #4a56e2;
-    transform: scale(1.02);
-  }
-}
-.buttonView {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-top: 20px;
-}
