+++ src/components/UserMenu/UserMenu.module.scss
@@ -15,6 +15,13 @@
   font-size: 18px;
   line-height: 1.5;
 }
+.ModalContent {
+  background-image: linear-gradient(to right, #e7eaf2 35%, #ffd8d0 80%);
+  padding: 30px;
+  width: 100%;
+  height: 100%;
+  border-radius: 20px;
+}
 .buttonLogout {
   display: flex;
   align-items: center;
@@ -36,6 +43,39 @@
   top: 25px;
   border: 1px solid #bdbdbd;
   margin-right: 15px;
+}
+.buttonLogoutYN {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100px;
+  height: 40px;
+  padding: 0px;
+  border: none;
+  border-radius: 8px;
+
+  background-color: #6e78e8;
+  cursor: pointer;
+  color: #ff6596;
+  font-size: 24px;
+  line-height: 1.33;
+  box-shadow: 0px 3px 1px -1px rgba(0, 0, 0, 0.2),
+    0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12);
+  transition: transform 150ms cubic-bezier(0.4, 0, 0.2, 1);
+  &:hover {
+    background-color: #4a56e2;
+    transform: scale(1.02);
+  }
+  &:focus {
+    background-color: #4a56e2;
+    transform: scale(1.02);
+  }
+}
+.buttonView {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-top: 20px;
 }
 @media screen and (max-width: 767px) {
   .name {
